+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changes
 * renamed Type.OSX to Type.macOS, you can use both 'OSX' and 'macOS' for the xcodebuild.type parameter
 * Introduced an entitlements parameter for signing that can contain a map with the entitlments settings that should be added
+* Added validation of signing certificate (if it is valid and has not expired): Issue #348
 
 # 0.14.5 (January 26, 2017)